Don't try to remove an empty file name.

+++ updater.php
@@ -351,6 +351,9 @@ class Updater
             );
             foreach ($obsoleteFiles as $file) {
                 $file = trim($file);
+                if ('' === $file) {
+                    continue;
+                }
                 if (!unlink($file)) {
                     $this->error("Could not remove '$file'");
                 }
